# lib/api-zod/src/generated/types/favorite.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Favorite {
  id: number;
  workId: string;
  title: string;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  createdAt: Date;
}