# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  /**
   * @minimum 1
   * @maximum 5
   */
  stars: number;
  /**
   * @minLength 1
   * @maxLength 80
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 4000
   */
  text: string;
}